Add OutputDeviceIcon component with dynamic icon rendering and namespace support
